Create diagnostic module for profiling and crash diagnostics
